rename pipe/pool to pipe/pooling
