+++ pooling/pooling_test.go
@@ -1,9 +1,10 @@
-package pool_test
+package pooling_test
 
 import (
 	"testing"
 
-	"pipelined.dev/pipe/pool"
+	"pipelined.dev/pipe/pooling"
+
 	"pipelined.dev/signal"
 )
 
@@ -13,14 +14,14 @@ func TestPool(t *testing.T) {
 		Length:   512,
 	}
 
-	p1 := pool.Get(alloc)
-	p2 := pool.Get(alloc)
+	p1 := pooling.Get(alloc)
+	p2 := pooling.Get(alloc)
 	if p1 != p2 {
 		t.Fatal("p1 must be equal to p2")
 	}
 
-	pool.Wipe()
-	p3 := pool.Get(alloc)
+	pooling.Wipe()
+	p3 := pooling.Get(alloc)
 	if p1 == p3 {
 		t.Fatal("p1 must not be equal to p3")
 	}
